ci: skip tarball GPG signing when no key is available

+++ ci/tarball.deploy.sh
@@ -3,4 +3,7 @@
 export TAG=$GITTAGNOV
 
 mv "ossia-score.tar.xz" "$BUILD_ARTIFACTSTAGINGDIRECTORY/ossia score-$TAG-src.tar.xz"
-mv "ossia-score.tar.xz.asc" "$BUILD_ARTIFACTSTAGINGDIRECTORY/ossia score-$TAG-src.tar.xz.asc"
+
+if [[ -f "ossia-score.tar.xz.asc" ]]; then
+  mv "ossia-score.tar.xz.asc" "$BUILD_ARTIFACTSTAGINGDIRECTORY/ossia score-$TAG-src.tar.xz.asc"
+fi
